tts-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
requests==2.32.3
openai-whisper==20240930
pyttsx3==2.98
